Update vscode.WSRMac
added ability to pop out the terminal. This is useful after first boot of computer and then opening up visual studio code and running caster from the terminal. This is because i like to pop out the terminal and put it onto the left monitor so it can function as a sort of gui when i am using caster